Rename SessionTime -> Clock.
Formatting.
