--- ninja-core/src/main/java/ninja/session/SessionTime.java
@@ -1,8 +1,0 @@
-package ninja.session;
-
-import com.google.inject.ImplementedBy;
-
-@ImplementedBy(SessionTimeImpl.class)
-public interface SessionTime {
-    long currentTimeMillis();
-}
