+++ ninja-core/src/main/java/ninja/utils/Clock.java
@@ -1,0 +1,8 @@
+package ninja.utils;
+
+import com.google.inject.ImplementedBy;
+
+@ImplementedBy(ClockImpl.class)
+public interface Clock {
+    long currentTimeMillis();
+}
